+++ NEWS.md
@@ -1,1 +1,7 @@
-There is currently no news.
+# wfr 0.5.0
+
+## new features
+
+* `saveOutput` function
+    * The default value of parameter `oFileName` is removed. Now its value must be provided.
+    * A new parameter `numberOutputFiles` is provided with default value `TRUE`. If `FALSE`, the output files will not be numbered.